Refactor files to use a different structure
