+++ src/base/strings.js
