+++ src/base/type-checker.js
