+++ src/generator/cli.js
@@ -19,7 +19,7 @@
 import fs from "fs";
 
 import { processFolder, processGitRepo } from "./processor";
-import "../templates";
+import "./templates";
 
 // Initially we will support only GitHub repos
 // TODO: support other urls. List can be found here:
