+++ src/generator/templates/asset.hbs
