+++ src/generator/templates/contract.hbs
