+++ src/generator/templates/index.js
