+++ src/generator/templates/package.hbs
