+++ src/generator/templates/script.hbs
