+++ src/generator/templates/version.hbs
